+++ libs/parser/test/integration/concurrency_test.go
@@ -34,12 +34,9 @@ func testConcurrentParses(t *testing.T, numGoroutines int, testName string) {
 	t.Helper()
 
 	// Sample Nomos source to parse
-	// Note: Updated to use inline references instead of deprecated top-level reference statements
 	source := `source:
   alias: myConfig
   type: yaml
-
-import:baseConfig:./base.csl
 
 database:
   host: localhost